[FLINK-33470] Moving (Join|Expand)TestPrograms and RestoreTests to the right package.
